requirements.txt
absl-py==0.2.0
anndata==0.6.20
appdirs==1.4.4
arviz==0.8.3
astor==0.6.2
autograd==1.3
autograd-gamma==0.4.1
backcall==0.1.0
beautifulsoup4==4.9.3
bioservices==1.7.11
bleach==1.5.0
certifi==2020.12.5
cffi==1.11.5
cftime==1.1.3
chardet==4.0.0
cloudpickle==0.5.3
colorama==0.4.4
colorlog==4.8.0
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
dill==0.2.7.1
easydev==0.11.0
entrypoints==0.2.3
future==0.17.1
gast==0.2.0
gevent==21.1.2
greenlet==1.0.0
grequests==0.6.0
grpcio==1.11.0
gseapy==0.10.4
h5py==2.7.1
html5lib==0.9999999
idna==2.10
imageio==2.9.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.10
joblib==0.12.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.6
kiwisolver==1.0.1
lifelines==0.22.3
llvmlite==0.31.0
lxml==4.6.3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.3
mistune==0.8.3
multiprocess==0.70.5
natsort==7.0.1
nbconvert==5.3.1
nbformat==4.4.0
netCDF4==1.5.3
networkx==2.1
notebook==5.6.0
numba==0.47.0
numexpr==2.7.1
numpy==1.14.3
opencv-python==3.4.0.12
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parse==1.8.2
parso==0.3.1
pathos==0.2.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
pkg-resources==0.0.0
plotly==4.3.0
pox==0.2.3
ppft==1.6.4.7.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.6.0
pyarrow==0.13.0
pycparser==2.18
Pygments==2.2.0
pymc3==3.6
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
pyvips==2.1.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.1
rdp==0.8
requests==2.25.1
requests-cache==0.5.2
retrying==1.3.3
rpy2==3.4.5
scanpy==1.4.3
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
soupsieve==2.1
statsmodels==0.11.1
suds-jurko==0.6
tables==3.6.1
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.4
toolz==0.9.0
tornado==5.1
tqdm==4.31.1
trackpy==0.4.2
traitlets==4.3.2
tzlocal==2.1
umap-learn==0.3.10
urllib3==1.26.4
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.1
wrapt==1.12.1
xarray==0.13.0
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.3.0
